rename(auth_style): 대문자로 변경
