+++ src/app/(route)/(auth)/(authForm)/find-pw/page.tsx
@@ -2,7 +2,7 @@
 
 import { useFormContext } from "react-hook-form";
 import { Button, InputText } from "@/components";
-import { ButtonStyle } from "../sign-up/_constant/authStyle";
+import { ButtonStyle } from "../sign-up/_constant/AUTH_STYLE";
 import { useRouter } from "next/navigation";
 import { cn } from "@/utils";
