+++ src/app/(route)/(auth)/(authForm)/sign-up/_constant/AUTH_STYLE.ts
